Refactor connection to evaluation service
