+++ src/main/java/org/damap/base/rest/EvaluationResource.java
@@ -1,18 +1,25 @@
-package org.damap.base.rest.evaluation;
+package org.damap.base.rest;
 
 import io.quarkus.security.Authenticated;
 import io.quarkus.security.AuthenticationFailedException;
 import io.quarkus.security.ForbiddenException;
 import jakarta.inject.Inject;
-import jakarta.ws.rs.*;
+import jakarta.ws.rs.GET;
+import jakarta.ws.rs.POST;
+import jakarta.ws.rs.Path;
+import jakarta.ws.rs.PathParam;
+import jakarta.ws.rs.Produces;
+import jakarta.ws.rs.QueryParam;
 import jakarta.ws.rs.core.MediaType;
 import java.util.List;
 import lombok.extern.jbosslog.JBossLog;
 import org.damap.base.rest.evaluation.dto.BenchmarkDTO;
 import org.damap.base.rest.evaluation.dto.EvaluationResultDTO;
+import org.damap.base.rest.evaluation.service.EvaluationService;
 import org.damap.base.security.SecurityService;
 import org.damap.base.validation.AccessValidator;
 
+/** EvaluationResource class. */
 @Path("/api/evaluation")
 @Authenticated
 @Produces(MediaType.APPLICATION_JSON)
@@ -25,6 +32,12 @@ public class EvaluationResource {
 
   @Inject EvaluationService evaluationService;
 
+  /**
+   * getBenchmarks.
+   *
+   * @return a {@link java.util.List} of {@link org.damap.base.rest.evaluation.dto.BenchmarkDTO}
+   *     objects
+   */
   @GET
   @Path("/benchmarks")
   public List<BenchmarkDTO> getBenchmarks() {
@@ -32,6 +45,14 @@ public class EvaluationResource {
     return evaluationService.listBenchmarks();
   }
 
+  /**
+   * assessBenchmark.
+   *
+   * @param dmpId a long
+   * @param benchmark a {@link java.lang.String} object
+   * @return a {@link java.util.List} of {@link
+   *     org.damap.base.rest.evaluation.dto.EvaluationResultDTO} objects
+   */
   @POST
   @Path("/assess/{dmpId}")
   public List<EvaluationResultDTO> assessBenchmark(
