+++ src/main/java/org/damap/base/rest/evaluation/dto/EvaluationMultipartBodyDTO.java
@@ -1,11 +1,12 @@
-package org.damap.base.rest.evaluation;
+package org.damap.base.rest.evaluation.dto;
 
 import jakarta.ws.rs.FormParam;
 import jakarta.ws.rs.core.MediaType;
 import java.io.InputStream;
 import org.jboss.resteasy.reactive.PartType;
 
-public class EvaluationMultipartBody {
+/** EvaluationMultipartBodyDTO class. */
+public class EvaluationMultipartBodyDTO {
 
   @FormParam("maDMP")
   @PartType(MediaType.APPLICATION_OCTET_STREAM)
