+++ src/main/java/org/damap/base/rest/evaluation/dto/EvaluationResultDTO.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import java.util.List;
 import lombok.Data;
 
+/** EvaluationResultDTO class. */
 @Data
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class EvaluationResultDTO {
